MET #20: first draft of simim/sat app refactor
